requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
pymongo>=4.5.0
pandas>=2.0.0
numpy>=1.24.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
python-dotenv>=1.0.0

